+++ client/src/io/worxbend/gitea4s/api/UsersApi.scala
@@ -2,7 +2,7 @@ package io.worxbend.gitea4s.api
 
 import io.worxbend.gitea4s.error.GiteaError
 import io.worxbend.gitea4s.http.UserSearchParams
-import io.worxbend.gitea4s.model.User
+import io.worxbend.gitea4s.model.{StopWatch, User}
 import zio.IO
 import zio.stream.ZStream
 
@@ -16,3 +16,5 @@ trait UsersApi:
   def following(username: String): ZStream[Any, GiteaError, User]
 
   def search(params: UserSearchParams): ZStream[Any, GiteaError, User]
+
+  def stopwatches: ZStream[Any, GiteaError, StopWatch]